requirements.txt
streamlit
langchain
langchain_community
langchain_core
langchain_groq
python-dotenv
pandas